+++ NEWS.md
@@ -1,4 +1,4 @@
-# NBDCtools 1.0.3
+# NBDCtools 1.1.0
 
 ## New Features
 
@@ -14,6 +14,10 @@
   the correct session IDs for that release. If this file is not found, users
   will get a warning message and futher validation is skipped. If the file is
   found but the session IDs do not match, an error is raised.
+- Added `add_custom_metadata()` function to allow users to add their own 
+  metadata to the NBDCtools environment. This is useful for users who have 
+  their own metadata. A new vignette "Add custom metadata" 
+  has been added to demonstrate how to use this function.
   
 ## Other Changes